requirements.txt
fastapi
uvicorn
reportlab
Pillow
python-multipart
gspread
google-auth-oauthlib
pandas